[refactor] mudança de nome do pacote + classe: vendas -> venda
