+++ src/main/java/br/com/spring/code/ecommerce/venda/Vendas.java
@@ -1,4 +1,4 @@
-package br.com.spring.code.ecommerce.vendas;
+package br.com.spring.code.ecommerce.venda;
 
 import java.util.List;
 
@@ -7,12 +7,20 @@ import br.com.spring.code.ecommerce.financeiro.Financeiro;
 import br.com.spring.code.ecommerce.financeiro.Pagamento;
 import br.com.spring.code.ecommerce.gestaopessoas.Pessoa;
 
+/**
+ * 
+ * @author thiagojacinto, carloshenrique e wilson
+ * 
+ * Definição: classe que representa os pagamentos do sistema de e-commerce.
+ *
+ */
 public class Vendas {	
+	
 	private Pessoa pessoa;
 	private List<Anuncio> anuncio;
 	private Pagamento pagamento;
 	private Financeiro financeiro;
-	private String tipoEnvio;
+	private String TIPO_DE_ENVIO;
 	private boolean confirmarVenda;
